+++ src/smallworldnet.jl
@@ -1,26 +1,26 @@
 
 ## El tipo SmallWorldNet con el constructor básico ----- ##
 type SmallWorldNet
-    L::Int64
-    Z::Int64
+    L::Int
+    Z::Int
     p::Float64
-    neighbours::Vector{Vector{Int64}}
+    neighbours::Vector{Vector{Int}}
 
-    function SmallWorldNet(L::Int64)
-        neighbours = Array(Vector{Int64},L)
+    function SmallWorldNet(L::Int)
+        neighbours = Array(Vector{Int},L)
         fill!(neighbours, [])
         new(L, 0, 0., neighbours)
     end
 end
-        
+
 import Base.show
- 
+
 show(io::IO, w::SmallWorldNet) = println(io, "L:$(w.L) Z:$(w.Z) p:$(w.p)")
 
 
 ## Las funciones de la red --------------------------- ##
 
-function HasNode(w::SmallWorldNet, n::Int64)
+function HasNode(w::SmallWorldNet, n::Int)
     n <= length(w.neighbours) ? true : false
 end
 
@@ -32,11 +32,11 @@ function GetNodes(w::SmallWorldNet)
     [1:length(w.neighbours)]
 end
 
-function GetNeighbours(w::SmallWorldNet, n::Int64)
+function GetNeighbours(w::SmallWorldNet, n::Int)
     w.neighbours[n]
 end
 
-function AddEdge!(w::SmallWorldNet, n1::Int64, n2::Int64)
+function AddEdge!(w::SmallWorldNet, n1::Int, n2::Int)
     if n1 == n2
         #print("Only one node given: ",n1)
         return
@@ -44,7 +44,7 @@ function AddEdge!(w::SmallWorldNet, n1::Int64, n2::Int64)
         #print("Edge already defined")
         return
     else
-        push!(w.neighbours[n1], n2) 
+        push!(w.neighbours[n1], n2)
         push!(w.neighbours[n2], n1)
     end
 end
@@ -53,20 +53,20 @@ end
 ## El constructor completo-------------------------- ##
 
 function AddNeighbourEdges!(w::SmallWorldNet)
-    modulo = Dict{Int64,Int64}()
+    modulo = Dict{Int,Int}()
     for i in 0:w.L-1
         modulo[-i] = w.L - i
         modulo[i+1] = i+1
         modulo[w.L+i] = i
     end
-    
+
     for n in 1:w.L
         rango = [n - div(w.Z,2):n-1, n+1:n + div(w.Z,2)]
         traducido = [modulo[t] for t in rango]
         for n2 in traducido
             AddEdge!(w, n, n2)
         end
-    end 
+    end
 end
 
 function AddRandomEdges!(w::SmallWorldNet)
@@ -79,13 +79,13 @@ function AddRandomEdges!(w::SmallWorldNet)
             cont += 1
         end
     end
-end     
-    
-function SmallWorldNet(L::Int64, Z::Int64, p::Float64)
+end
+
+function SmallWorldNet(L::Int, Z::Int, p::Float64)
     w = SmallWorldNet(L)
     w.Z = Z
     w.p = p
-    
+
     AddNeighbourEdges!(w)
     AddRandomEdges!(w)
 
@@ -99,5 +99,5 @@ function SmallWorldNetWithNoStep(w::SmallWorldNet)
     for i in 1:w.L
         wc.neighbours[i] = push!([copy(w.neighbours[i])...],i)
     end
-    return wc 
+    return wc
 end
